dbackend/requirements.txt
alabaster==0.7.16
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docutils==0.20.1
drf-yasg==1.21.7
filelock==3.15.4
fsspec==2024.9.0
huggingface-hub==0.24.6
hyperlink==21.0.0
idna==3.8
imagesize==1.4.1
implicit==0.7.2
incremental==24.7.2
inflection==0.5.1
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.4.4
scipy==1.14.1
service-identity==24.1.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
sympy==1.13.2
tbb==2021.13.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
Twisted==24.7.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
zope.interface==7.0.3
